test(runtime): isolate filesystem and SYS probes
